Cast strlen to int for printf, so we're 64-bit clean.

+++ fc-glyphname/fc-glyphname.c
@@ -286,7 +286,7 @@ main (int argc, char **argv)
     for (i = 0; i < nraw; i++)
 	printf ("static struct { FcChar32 ucs; FcChar8 name[%d]; }"
 	        " glyph%d = { 0x%lx, \"%s\" };\n",
-	        strlen (raw[i]->name) + 1,
+	        (int) strlen (raw[i]->name) + 1,
 		i, (unsigned long) raw[i]->ucs, raw[i]->name);
 
     /*
